Refactored Brute Force Protection
CVE Score of 5.3 (Medium).

Brute force protection on the `!verify` command was incorrectly implemented . The user would be blocked if the value `verify_attempts >= 5`. However, this value was only ever increased AFTER the successful verification phase. Meaning brute force was not limited.

The solution was to move the `verify_attempts` checker outside of the `if token_list[ctx.author.id] == arg:` statement